# services/billing/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Billing Service API for OCR passport scanning service. Manages accounts, balances, subscriptions, and payments.",
        "title": "Billing Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@api-scan.example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/accounts": {
            "post": {
                "description": "Create a new billing account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create new account",
                "responses": {
                    "201": {
                        "description": "Created account",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/balance": {
            "get": {
                "description": "Get real and prepaid balance for an account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account balance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/payments": {
            "post": {
                "description": "Create a payment order for balance top-up",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/reserve": {
            "post": {
                "description": "Reserve funds for an operation (two-phase commit)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Reserve funds",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reservation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ReserveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ReserveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Insufficient balance",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/subscriptions": {
            "post": {
                "description": "Create a new subscription for an account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Create subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subscription request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Insufficient balance",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/subscriptions/upgrade": {
            "post": {
                "description": "Upgrade subscription to a higher tier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Upgrade subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Upgrade request {\\",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the billing service is running",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/payments/{id}": {
            "get": {
                "description": "Get payment order details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}/commit": {
            "post": {
                "description": "Commit a reserved transaction (two-phase commit)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Commit transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID (request_id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}/rollback": {
            "post": {
                "description": "Rollback a reserved transaction (two-phase commit)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Rollback transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID (request_id)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.BalanceResponse": {
            "description": "Account balance with real and prepaid amounts",
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer",
                    "example": 123
                },
                "prepaid_balance_rub": {
                    "type": "number",
                    "example": 5000
                },
                "real_balance_rub": {
                    "type": "number",
                    "example": 15000.5
                }
            }
        },
        "handler.CreatePaymentRequest": {
            "description": "Request to create a payment order",
            "type": "object",
            "properties": {
                "amount_rub": {
                    "type": "number",
                    "example": 1000
                },
                "description": {
                    "type": "string",
                    "example": "Balance top-up"
                },
                "return_url": {
                    "type": "string",
                    "example": "https://example.com/payment/success"
                }
            }
        },
        "handler.CreatePaymentResponse": {
            "description": "Payment order created successfully",
            "type": "object",
            "properties": {
                "amount_rub": {
                    "type": "number",
                    "example": 1000
                },
                "order_id": {
                    "type": "integer",
                    "example": 42
                },
                "payment_url": {
                    "type": "string",
                    "example": "https://yookassa.ru/payments/mock_42_123456"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                }
            }
        },
        "handler.CreateSubscriptionRequest": {
            "description": "Request to create a new subscription",
            "type": "object",
            "properties": {
                "payment_method": {
                    "type": "string",
                    "example": "balance"
                },
                "tariff_code": {
                    "type": "string",
                    "example": "pro"
                }
            }
        },
        "handler.CreateSubscriptionResponse": {
            "description": "Subscription created successfully",
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2026-04-18T10:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "prepaid_remaining_rub": {
                    "type": "number",
                    "example": 6000
                },
                "started_at": {
                    "type": "string",
                    "example": "2026-03-18T10:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "tariff_code": {
                    "type": "string",
                    "example": "pro"
                }
            }
        },
        "handler.ErrorResponse": {
            "description": "Error response with code and message",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "PAYMENT_REQUIRED"
                },
                "error": {
                    "type": "string",
                    "example": "insufficient balance"
                }
            }
        },
        "handler.ReserveRequest": {
            "description": "Request to reserve funds for an operation",
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string",
                    "example": "req_abc123"
                },
                "service_id": {
                    "type": "string",
                    "example": "passport_rf"
                }
            }
        },
        "handler.ReserveResponse": {
            "description": "Reservation result",
            "type": "object",
            "properties": {
                "amount_rub": {
                    "type": "number",
                    "example": 7
                },
                "charge_type": {
                    "type": "string",
                    "example": "prepaid"
                },
                "reserved": {
                    "type": "boolean",
                    "example": true
                },
                "transaction_id": {
                    "type": "string",
                    "example": "req_abc123"
                }
            }
        },
        "handler.SuccessResponse": {
            "description": "Success response",
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        }
    }
}